fix bugs, write tests, organise files

+++ src/types.rs
@@ -1,4 +1,4 @@
 pub type ClientId = u16;
 pub type TxId = u32;
 // TODO: 4dp
-pub type Balance = f32;
+pub type Balance = f64;
